Hotels klaar
Hotel zijn zo goed als klaar. De migrations werken weer niet. Ik heb constant problemen door de PascalMigration. Vorige keer ook al

+++ MobyPark/DTOs/Hotel/HotelPassDtos.cs
@@ -19,7 +19,7 @@ public class AdminCreateHotelPassDto
     [Required]
     public string LicensePlate { get; set; }
     [Required]
-    public int ParkingLotId { get; set; }
+    public long ParkingLotId { get; set; }
     [Required]
     public DateTime Start { get; set; }
     [Required]
